Run go mod tidy as part of GitHub actions
Make sure go.mod and go.sum match the current source in the module.

Running `go mod tidy` also removes the indirect dependency on
github.com/dvyukov/go-fuzz which was probably added when running the `go
get` commands mentioned in the README to install the `go-fuzz` and
`go-fuzz-build` in the package directory.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>