netaddr: add functionality to IPRange for feature parity
Add MustParseIPRange, IPRange.{IsZero,AppendTo,MarshalText,UnmarshalText}
to bring it to feature parity with the other common types.

Similar to the MarshalText methods of the other types,
it never returns an error for invalid values.
This behavior should probably change, but should be consistently
done for all other types (see #169).

Fixes #201

Signed-off-by: Joe Tsai <joetsai@digital-static.net>