Add dotted decimal output for mapped addr
For IPv4 mapped IPv6 addresses it is commont to see the IPv4 part of the
address as dotted decmal after a prefix of `::ffff:` making
`::ffff:c000:0202` be printed as `::ffff:192.0.2.2`.

This commit implements a method that prints this format for mapped
addresses, and uses `String()` for all non mapped addresses.